ship_tracking/risk_metric/requirements.txt
google-cloud-bigquery
supabase
python-dotenv
openai>=1.0.0
